scripts/requirements.txt
gpxpy
exif
pandas
geopy
Pillow
python-dateutil
